Fix heap loop and dig_serial
The process loop missed a .items(). It has to continue on if there are
no heaps within the dump (which happens if the noise diodes only
start up in the middle of a capture block). Only save the dig_serial
if it has something, since it is immutable and we only have one shot.