New --capture-heaps option
This captures some of the underlying heap data to telstate to aid
debugging of the algorithms generating the ECP64 products. More
specifically, it stores the relevant parts of the heaps that make up
each L0 dump product, with the same timestamp as the corresponding
L0 dump. The data, nd_on flags and n_accs are stored per receptor per
stream (hh, vv, hv). The script option takes a comma-separated list
of receptor names to identify the streams to capture.