Create SPEAD descriptors locally
Go back to the original hacky `ecp64.py` approach :-) Just add n_accs,
which was added in the meantime.

This is because SPEAD descriptors might be part of the network problems
encountered with the spectrometers (given that they all emanate from a
single server). Then again, it might make no difference as well.